extract validated form element to its own component and replace instances thereof
