+++ app/routes/shared/components/VorabcheckPage.tsx
@@ -1,14 +1,9 @@
-import { useLoaderData, useLocation, useParams } from "@remix-run/react";
-import { ValidatedForm } from "remix-validated-form";
+import { useLoaderData } from "@remix-run/react";
 import Background from "~/components/Background";
 import Container from "~/components/Container";
-import { ButtonNavigation } from "~/components/form/ButtonNavigation";
 import { ProgressBar } from "~/components/form/ProgressBar";
+import ValidatedFlowForm from "~/components/form/ValidatedFlowForm";
 import PageContent from "~/components/PageContent";
-import { StrapiFormComponents } from "~/services/cms/components/StrapiFormComponents";
-import { splatFromParams } from "~/services/params";
-import { CSRFKey } from "~/services/security/csrf/csrfKey";
-import { validatorForFieldnames } from "~/services/validation/buildStepValidator";
 import type { loader } from "../vorabcheck.server";
 
 export function VorabcheckPage() {
@@ -20,11 +15,6 @@ export function VorabcheckPage() {
     progressProps,
     buttonNavigationProps,
   } = useLoaderData<typeof loader>();
-  const stepId = splatFromParams(useParams());
-  const { pathname } = useLocation();
-  const fieldNames = formElements.map((entry) => entry.name);
-  const validator = validatorForFieldnames(fieldNames, pathname);
-
   return (
     <Background backgroundColor="blue">
       <div className="min-h-screen">
@@ -37,22 +27,12 @@ export function VorabcheckPage() {
                 className="ds-stack-16"
                 fullScreen={false}
               />
-              <ValidatedForm
-                id={`${stepId}_form`}
-                method="post"
-                validator={validator}
-                defaultValues={stepData}
-                noValidate
-                action={pathname}
-              >
-                <input type="hidden" name={CSRFKey} value={csrf} />
-                <div className="ds-stack-40">
-                  <div className="ds-stack-40">
-                    <StrapiFormComponents components={formElements} />
-                  </div>
-                  <ButtonNavigation {...buttonNavigationProps} />
-                </div>
-              </ValidatedForm>
+              <ValidatedFlowForm
+                stepData={stepData}
+                csrf={csrf}
+                formElements={formElements}
+                buttonNavigationProps={buttonNavigationProps}
+              />
             </div>
           </div>
         </Container>
